requirements.txt
Flask
pycryptodomex
pyinstaller
requests